+++ chords_serial.py
